# schemas/api-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wasm.fyi benchmark API",
    "version": "2.0.0",
    "description": "Current benchmark measurements with indexed cursor pagination and atomic capture publication."
  },
  "paths": {
    "/api/platforms": {
      "get": {
        "summary": "List measured platforms",
        "responses": {
          "200": {
            "description": "Platform catalog at the current revision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/PlatformList"
                }
              }
            }
          },
          "429": {
            "description": "Request budget exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/benchmarks": {
      "get": {
        "summary": "Page current benchmark results for one platform and phase",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "schema": {
              "$ref": "benchmark.schema.json#/$defs/Digest"
            },
            "description": "Required on the first page; carried by cursor thereafter"
          },
          {
            "name": "phase",
            "in": "query",
            "schema": {
              "enum": [
                "compile",
                "instantiate",
                "first-call",
                "steady"
              ],
              "default": "steady"
            },
            "description": "Must match cursor when supplied"
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 4096
            },
            "description": "Opaque signed continuation; binds platform, phase and revision"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 200
            },
            "description": "Maximum rows; may change between pages"
          }
        ],
        "responses": {
          "200": {
            "description": "Current page; nextCursor is null on the last page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/BenchmarkPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query or cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Unknown platform",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Revision changed; restart pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Request budget exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/captures": {
      "post": {
        "summary": "Atomically upsert a validated capture",
        "security": [
          {
            "publisherToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "benchmark.schema.json#/$defs/Capture"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Durable acknowledgment; equal-time and older cells are ignored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/PublicationReceipt"
                }
              }
            }
          },
          "400": {
            "description": "Invalid capture",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Publisher token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Capture exceeds 1 MiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Use application/json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Request budget exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Read-only mode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/healthz": {
      "get": {
        "summary": "Database readiness",
        "responses": {
          "200": {
            "description": "Ready",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ready": {
                      "const": true
                    }
                  },
                  "required": [
                    "ready"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/history": {
      "get": {
        "summary": "Paginated retained source versions for one platform and phase",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "schema": {
              "$ref": "benchmark.schema.json#/$defs/Digest"
            },
            "description": "Required on the first page; carried by cursor thereafter"
          },
          {
            "name": "phase",
            "in": "query",
            "schema": {
              "enum": [
                "compile",
                "instantiate",
                "first-call",
                "steady"
              ],
              "default": "steady"
            },
            "description": "Must match cursor when supplied"
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 4096
            },
            "description": "Opaque signed continuation; binds platform, phase and revision"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 200
            },
            "description": "Maximum rows; may change between pages"
          }
        ],
        "responses": {
          "200": {
            "description": "Current page; nextCursor is null on the last page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/BenchmarkPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query or cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Unknown platform",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Revision changed; restart pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Request budget exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "benchmark.schema.json#/$defs/Problem"
                }
              }
            }
          }
        },
        "operationId": "history"
      }
    }
  },
  "components": {
    "securitySchemes": {
      "publisherToken": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {}
  }
}